+++ tests/test_draw.py
@@ -28,23 +28,22 @@ def minuit(request):
 
 
 def test_profile(minuit):
-    minuit.draw_profile("x")
-    minuit.minos("x")
-    minuit.draw_profile("x")
+    minuit.draw_profile("x")  # plots with hesse errors
+    minuit.minos()
+    minuit.draw_profile("x")  # plots with minos errors
 
 
 def test_mnprofile(minuit):
-    minuit.minos("x")
-    minuit.draw_mnprofile("x")
+    minuit.draw_mnprofile("x")  # plots with hesse errors
+    minuit.minos()
+    minuit.draw_mnprofile("x")  # plots with minos errors
 
 
 def test_mncontour(minuit):
-    minuit.minos()
     minuit.draw_mncontour("x", "y")
 
 
 def test_drawcontour(minuit):
-    minuit.minos()
     minuit.draw_contour("x", "y")
-    minuit.draw_contour("x", "x", bins=20, bound=2)
-    minuit.draw_contour("x", "x", bins=20, bound=((-10, 10), (-10, 10)))
+    minuit.draw_contour("x", "x", size=20, bound=2)
+    minuit.draw_contour("x", "x", size=20, bound=((-10, 10), (-10, 10)))